# БР 2.1/Samsonov_Alexander/labs/lab1/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Ingredient": {
				"properties": {
					"quantity": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"quantity",
					"name"
				],
				"type": "object"
			},
			"RecipeContentBlock": {
				"anyOf": [
					{
						"properties": {
							"content": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"text"
								],
								"nullable": false
							}
						},
						"required": [
							"content",
							"type"
						],
						"type": "object"
					},
					{
						"properties": {
							"alt": {
								"type": "string"
							},
							"url": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"image"
								],
								"nullable": false
							}
						},
						"required": [
							"url",
							"type"
						],
						"type": "object"
					}
				]
			},
			"Recipe": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/Ingredient"
						},
						"type": "array"
					},
					"content": {
						"items": {
							"$ref": "#/components/schemas/RecipeContentBlock"
						},
						"type": "array"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array"
					},
					"likes": {
						"items": {
							"$ref": "#/components/schemas/Like"
						},
						"type": "array"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"ingredients",
					"content",
					"author",
					"tags",
					"likes",
					"comments",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"description": "User model",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "The unique identifier for the user",
						"example": 1
					},
					"name": {
						"type": "string",
						"description": "The user's full name",
						"example": "John Doe"
					},
					"email": {
						"type": "string",
						"description": "The user's email address",
						"example": "john.doe@example.com"
					},
					"password": {
						"type": "string",
						"description": "The user's password (hashed)",
						"example": "$2b$10$..."
					},
					"recipes": {
						"items": {
							"$ref": "#/components/schemas/Recipe"
						},
						"type": "array",
						"description": "Recipes created by the user"
					},
					"likes": {
						"items": {
							"$ref": "#/components/schemas/Like"
						},
						"type": "array",
						"description": "Likes given by the user"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array",
						"description": "Comments made by the user"
					},
					"subscriptions": {
						"items": {
							"$ref": "#/components/schemas/Subscription"
						},
						"type": "array",
						"description": "Users that this user is subscribed to"
					},
					"subscribers": {
						"items": {
							"$ref": "#/components/schemas/Subscription"
						},
						"type": "array",
						"description": "Users that are subscribed to this user"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"recipes",
					"likes",
					"comments",
					"subscriptions",
					"subscribers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Like": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"recipe": {
						"$ref": "#/components/schemas/Recipe"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user",
					"recipe",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"recipe": {
						"$ref": "#/components/schemas/Recipe"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"content",
					"user",
					"recipe",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Subscription": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"subscriber": {
						"$ref": "#/components/schemas/User"
					},
					"creator": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"subscriber",
					"creator",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tag": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"recipes": {
						"items": {
							"$ref": "#/components/schemas/Recipe"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"recipes"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_Tag_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"recipes": {
						"items": {
							"$ref": "#/components/schemas/Recipe"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "lab1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/email/{email}": {
			"get": {
				"operationId": "GetUserByEmail",
				"responses": {
					"200": {
						"description": "The user with the specified email",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Get a user by email",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "The email of the user to retrieve",
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "john.doe@example.com"
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "The user with the specified ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Get a user by ID",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the user to retrieve",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": 1
					}
				]
			}
		},
		"/tags": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tag"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all tags",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					}
				},
				"description": "Create a new tag",
				"tags": [
					"Tags"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Tag_"
							}
						}
					}
				}
			}
		},
		"/tags/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					}
				},
				"description": "Get a tag by ID",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					}
				},
				"description": "Update an existing tag",
				"tags": [
					"Tags"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Tag_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Delete a tag",
				"tags": [
					"Tags"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tags/name/{name}": {
			"get": {
				"operationId": "GetTagByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					}
				},
				"description": "Get a tag by name",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/recipes/tag/{tagName}": {
			"get": {
				"operationId": "GetRecipesByTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Recipe"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get recipes by tag name",
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tagName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/likes/user/{userId}": {
			"get": {
				"operationId": "GetLikesByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Like"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get likes by user ID",
				"tags": [
					"Likes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/comments/user/{userId}": {
			"get": {
				"operationId": "GetCommentsByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Comment"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get comments by user ID",
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}